requirements.txt
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
fastapi==0.111.0
python-telegram-bot==21.1.1
pydantic-settings==2.2.1
#Pillow==10.4.0
